+++ docs/source/user_guide/changelog.rst
@@ -13,6 +13,7 @@ Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 - Fix bug occurring when session folder contains non-ASCII characters.
 - Another fix to "Too many open files" error by disabling ``pin_memory`` in Protocol 3 pre-training.
 - Fix the automatic gap detector in Validator's interpolator when starting the interpolation by double clicking on a centroid.
+- Fix individual videos cropping to keep the centroid of the animal in the center of the video when some padding has to be added.
 
 5.2.11
 ======